Module epic_p2p::types

Structs

Capabilities

Options for what type of interaction a peer supports

P2PConfig

Configuration for the peer-to-peer server.

PeerAddr

PeerInfo

General information about a connected peer that’s useful to other modules.

PeerInfoDisplay

Flatten out a PeerInfo and nested PeerLiveInfo (taking a read lock on it) so we can serialize/deserialize the data for the API and the TUI.

PeerLiveInfo

TxHashSetRead

The full txhashset data along with indexes required for a consumer to rewind to a consistent requested state.

Enums

Direction

Error

ReasonForBan

Seeding

Type of seeding the server will use to find other peers on the network.

Constants

MAX_BLOCK_BODIES

Maximum number of block bodies a peer should ever ask for and send

MAX_BLOCK_HEADERS

Maximum number of block headers a peer should ever send

MAX_LOCATORS

Maximum number of block header hashes to send as part of a locator

MAX_PEER_ADDRS

Maximum number of peer addresses a peer should ever send

Traits

ChainAdapter

Bridge between the networking layer and the rest of the system. Handles the forwarding or querying of blocks and transactions from the network among other things.

NetAdapter

Additional methods required by the protocol that don’t need to be externally implemented.