Trait epic_p2p::types::ChainAdapter

pub trait ChainAdapter: Sync + Send {

    fn total_difficulty(&self) -> Result<Difficulty, Error>;

    fn total_height(&self) -> Result<u64, Error>;

    fn transaction_received(
        &self, 
        tx: Transaction, 
        stem: bool
    ) -> Result<bool, Error>;

    fn get_transaction(&self, kernel_hash: Hash) -> Option<Transaction>;

    fn tx_kernel_received(
        &self, 
        kernel_hash: Hash, 
        peer_info: &PeerInfo
    ) -> Result<bool, Error>;

    fn block_received(
        &self, 
        b: Block, 
        peer_info: &PeerInfo, 
        opts: Options
    ) -> Result<bool, Error>;

    fn compact_block_received(
        &self, 
        cb: CompactBlock, 
        peer_info: &PeerInfo
    ) -> Result<bool, Error>;

    fn header_received(
        &self, 
        bh: BlockHeader, 
        peer_info: &PeerInfo
    ) -> Result<bool, Error>;

    fn headers_received(
        &self, 
        bh: &[BlockHeader], 
        peer_info: &PeerInfo
    ) -> Result<bool, Error>;

    fn locate_headers(
        &self, 
        locator: &[Hash]
    ) -> Result<Vec<BlockHeader>, Error>;

    fn get_block(&self, h: Hash) -> Option<Block>;

    fn kernel_data_read(&self) -> Result<File, Error>;

    fn kernel_data_write(&self, reader: &mut dyn Read) -> Result<bool, Error>;

    fn txhashset_read(&self, h: Hash) -> Option<TxHashSetRead>;

    fn txhashset_archive_header(&self) -> Result<BlockHeader, Error>;

    fn txhashset_receive_ready(&self) -> bool;

    fn txhashset_download_update(
        &self, 
        start_time: DateTime<Utc>, 
        downloaded_size: u64, 
        total_size: u64
    ) -> bool;

    fn txhashset_write(
        &self, 
        h: Hash, 
        txhashset_data: File, 
        peer_peer_info: &PeerInfo
    ) -> Result<bool, Error>;

    fn get_tmp_dir(&self) -> PathBuf;

    fn get_tmpfile_pathname(&self, tmpfile_name: String) -> PathBuf;

}

Bridge between the networking layer and the rest of the system. Handles the forwarding or querying of blocks and transactions from the network among other things.

Required methods

fn total_difficulty(&self) -> Result<Difficulty, Error>

Current total difficulty on our chain

fn total_height(&self) -> Result<u64, Error>

Current total height

fn transaction_received(
&self,
tx: Transaction,
stem: bool
) -> Result<bool, Error>

A valid transaction has been received from one of our peers

fn get_transaction(&self, kernel_hash: Hash) -> Option<Transaction>

fn tx_kernel_received(
&self,
kernel_hash: Hash,
peer_info: &PeerInfo
) -> Result<bool, Error>

fn block_received(
&self,
b: Block,
peer_info: &PeerInfo,
opts: Options
) -> Result<bool, Error>

A block has been received from one of our peers. Returns true if the block could be handled properly and is not deemed defective by the chain. Returning false means the block will never be valid and may result in the peer being banned.

fn compact_block_received(
&self,
cb: CompactBlock,
peer_info: &PeerInfo
) -> Result<bool, Error>

fn header_received(
&self,
bh: BlockHeader,
peer_info: &PeerInfo
) -> Result<bool, Error>

fn headers_received(
&self,
bh: &[BlockHeader],
peer_info: &PeerInfo
) -> Result<bool, Error>

A set of block header has been received, typically in response to a block header request.

fn locate_headers(&self, locator: &[Hash]) -> Result<Vec<BlockHeader>, Error>

Finds a list of block headers based on the provided locator. Tries to identify the common chain and gets the headers that follow it immediately.

fn get_block(&self, h: Hash) -> Option<Block>

Gets a full block by its hash.

fn kernel_data_read(&self) -> Result<File, Error>

fn kernel_data_write(&self, reader: &mut dyn Read) -> Result<bool, Error>

fn txhashset_read(&self, h: Hash) -> Option<TxHashSetRead>

Provides a reading view into the current txhashset state as well as the required indexes for a consumer to rewind to a consistant state at the provided block hash.

fn txhashset_archive_header(&self) -> Result<BlockHeader, Error>

Header of the txhashset archive currently being served to peers.

fn txhashset_receive_ready(&self) -> bool

Whether the node is ready to accept a new txhashset. If this isn’t the case, the archive is provided without being requested and likely an attack attempt. This should be checked before downloading the whole state data.

fn txhashset_download_update(
&self,
start_time: DateTime<Utc>,
downloaded_size: u64,
total_size: u64
) -> bool

Update txhashset downloading progress

fn txhashset_write(
&self,
h: Hash,
txhashset_data: File,
peer_peer_info: &PeerInfo
) -> Result<bool, Error>

Writes a reading view on a txhashset state that’s been provided to us. If we’re willing to accept that new state, the data stream will be read as a zip file, unzipped and the resulting state files should be rewound to the provided indexes.

fn get_tmp_dir(&self) -> PathBuf

Get the Epic specific tmp dir

fn get_tmpfile_pathname(&self, tmpfile_name: String) -> PathBuf

Get a tmp file path in above specific tmp dir (create tmp dir if not exist) Delete file if tmp file already exists