Trait epic_p2p::types::NetAdapter

pub trait NetAdapter: ChainAdapter {
    fn find_peer_addrs(&self, capab: Capabilities) -> Vec<PeerAddr>;

    fn peer_addrs_received(&self, _: Vec<PeerAddr>);

    fn peer_difficulty(&self, _: PeerAddr, _: Difficulty, _: u64, _: i64);

    fn is_banned(&self, addr: PeerAddr) -> bool;
}

Additional methods required by the protocol that don’t need to be externally implemented.

Required methods

fn find_peer_addrs(&self, capab: Capabilities) -> Vec<PeerAddr>

Find good peers we know with the provided capability and return their addresses.

fn peer_addrs_received(&self, _: Vec<PeerAddr>)

A list of peers has been received from one of our peers.

fn peer_difficulty(&self, _: PeerAddr, _: Difficulty, _: u64, _: i64)

Heard total_difficulty from a connected peer (via ping/pong).

fn is_banned(&self, addr: PeerAddr) -> bool

Is this peer currently banned?